fix deprecation warnings for wx2.9 (IMPORTANT: this change may have introduced some bugs)

+++ trunk/wxCode/components/keybinder/src/keybinder.cpp
@@ -556,7 +556,7 @@ bool wxCmd::Load(wxConfigBase *p, const wxString &key)
 // wxCmdArray
 // --------------------
 
-void wxCmdArray::Remove(int n)
+void wxCmdArray::Remove(size_t n)
 {
     if (n < 0 || n >= GetCount())
         return;
@@ -570,7 +570,7 @@ void wxCmdArray::Remove(int n)
 
 void wxCmdArray::Clear()
 {
-    for (int i=GetCount(); i > 0; i--)
+    for (size_t i=GetCount(); i > 0; i--)
         Remove(0);
 
     // the array should be already empty
@@ -745,7 +745,8 @@ void wxKeyBinder::Detach(wxWindow *p)
 
 void wxKeyBinder::DetachAll()
 {
-    wxKBLogDebug(wxT("wxKeyBinder::DetachAll - detaching from all my [%d] targets"), GetAttachedWndCount());
+    wxKBLogDebug(wxT("wxKeyBinder::DetachAll - detaching from all my [%d] targets"), 
+				 (int)GetAttachedWndCount());
 
     // delete all handlers (they will automatically remove themselves from
     // event handler chains)
@@ -837,7 +838,7 @@ bool wxKeyBinder::Save(wxConfigBase *cfg, const wxString &key, bool bCleanOld) c
     if (bCleanOld && cfg->Exists(basekey))
         cfg->DeleteGroup(basekey);      // delete old stuff...
 
-    for (int i=0; i < m_arrCmd.GetCount(); i++) {
+    for (size_t i=0; i < m_arrCmd.GetCount(); i++) {
 
         wxCmd *curr = m_arrCmd.Item(i);
 
@@ -976,11 +977,11 @@ bool wxKeyProfileArray::Save(wxConfigBase *cfg, const wxString &key, bool bClean
     if (!cfg->Write(basekey + wxT("nSelProfile"), m_nSelected))
         return FALSE;
 
-    for (int i=0; i<GetCount(); i++)
+    for (size_t i=0; i<GetCount(); i++)
 
         // save all our elements into a subkey of the given key
         b &= Item(i)->Save(cfg, basekey + wxKEYPROFILE_CONFIG_PREFIX +
-                                    wxString::Format(wxT("%d"), i), bCleanOld);
+                                    wxString::Format(wxT("%d"), (int)i), bCleanOld);
 
     // if required, remove any previously stored key profile...
     if (bCleanOld) {
@@ -1000,7 +1001,7 @@ bool wxKeyProfileArray::Save(wxConfigBase *cfg, const wxString &key, bool bClean
                 wxString id=str.Right(str.Len()-wxString(wxKEYPROFILE_CONFIG_PREFIX).Len());
                 id.ToLong(&n);
 
-                if (n >= GetCount()) {
+                if (n >= (long)GetCount()) {
 
                     // this is a profile which was saved in a previous session
                     // but which has now been removed by the user... remove it
@@ -1134,8 +1135,8 @@ wxKeyConfigPanel::~wxKeyConfigPanel()
 {
     // with the AddXXXXX functions we created wxKeyProfiles which we
     // then added into the m_pKeyProfiles combobox... we now must delete them.
-    for (int i=0; i < m_pKeyProfiles->GetCount(); i++) {
-        wxKeyProfile *data = (wxKeyProfile *)m_pKeyProfiles->GetClientData(i);
+    for (size_t i=0; i < m_pKeyProfiles->GetCount(); i++) {
+        wxKeyProfile *data = (wxKeyProfile *)m_pKeyProfiles->GetClientData((unsigned int)i);
 
         // we can delete the client data safely because wxComboBox will leave
         // the client data field untouched...
@@ -1369,7 +1370,7 @@ void wxKeyConfigPanel::ImportKeyProfileCmd(const wxKeyProfile &toimport,
     } else {
 
         const wxCmdArray *arr = toimport.GetArray();
-        for (int i=0; i < (int)arr->GetCount(); i++) {
+        for (size_t i=0; i < arr->GetCount(); i++) {
 
             // create a list of items containing as untyped client data
             // (void*) the INT which is their ID...
@@ -1404,7 +1405,7 @@ void wxKeyConfigPanel::AddProfile(const wxKeyProfile &p)
 void wxKeyConfigPanel::AddProfiles(const wxKeyProfileArray &arr)
 {
     // copy the given profiles into the listbox data list
-    for (int i=0; i < arr.GetCount(); i++) {
+    for (size_t i=0; i < arr.GetCount(); i++) {
         wxKeyProfile *copy = new wxKeyProfile(*arr.Item(i));
         m_pKeyProfiles->Append(arr.Item(i)->GetName(), (void *)copy);
     }
@@ -1414,7 +1415,8 @@ void wxKeyConfigPanel::AddProfiles(const wxKeyProfileArray &arr)
 
 void wxKeyConfigPanel::SetSelProfile(int n)
 {
-    wxASSERT(m_pKeyProfiles && n >= 0 && n < m_pKeyProfiles->GetCount());
+    wxASSERT(m_pKeyProfiles && n >= 0 && 
+			 n < (int)m_pKeyProfiles->GetCount());
 
     m_pKeyProfiles->SetSelection(n);
     m_nCurrentProf = n;
@@ -1481,7 +1483,7 @@ wxCmd *wxKeyConfigPanel::GetSelCmd() const
         if (sel < 0)
             return NULL;
 
-        id = (long)m_pCommandsList->GetClientData(sel);
+        id = (long)(m_pCommandsList->GetClientData(sel));
     }
 
     return m_kBinder.GetCmd(id);
@@ -1538,7 +1540,7 @@ wxKeyProfileArray wxKeyConfigPanel::GetProfiles() const
     // NB: it's very important to *copy* the profiles into the new array
     //     since the ddestructor of wxKeyConfigPanel expect the m_pKeyProfiles
     //     control to contain always valid pointers NOT shared with anyone else
-    for (int i=0; i<m_pKeyProfiles->GetCount(); i++)
+    for (size_t i=0; i<m_pKeyProfiles->GetCount(); i++)
         arr.Add(new wxKeyProfile(*GetProfile(i)));
     arr.SetSelProfile(GetSelProfileIdx());
 
@@ -1798,8 +1800,8 @@ void wxKeyConfigPanel::OnCategorySelected(wxCommandEvent &ev)
 
     // clear the old elements & insert the new ones
     m_pCommandsList->Clear();
-    for (int i=0; i < (int)arr.GetCount(); i++)
-        m_pCommandsList->Append(arr.Item(i), (void *)data->GetID(i));
+    for (size_t i=0; i < arr.GetCount(); i++)
+        m_pCommandsList->Append(arr.Item(i), (void *)data->GetID((int)i));
 
     // select the first
     m_pCommandsList->Select(0);
@@ -1972,7 +1974,7 @@ void wxKeyConfigPanel::OnAddProfile(wxCommandEvent &)
 
         // if the name is the same of one of the existing profiles, we have to abort...
         valid = TRUE;
-        for (int j=0; j < m_pKeyProfiles->GetCount(); j++)
+        for (size_t j=0; j < m_pKeyProfiles->GetCount(); j++)
             valid &= (GetProfile(j)->GetName() != dlg.GetValue());
 
         if (!valid) {
@@ -2010,7 +2012,7 @@ void wxKeyConfigPanel::OnRemoveProfile(wxCommandEvent &)
     // update the currently selected profile
     int newsel = m_nCurrentProf-1;
     if (newsel < 0) newsel=0;
-    wxASSERT(newsel < m_pKeyProfiles->GetCount());
+    wxASSERT(newsel < (int)m_pKeyProfiles->GetCount());
 
     // keep sync m_nCurrentProf with the currently really selected item
     SetSelProfile(newsel);
